fix(codemap): satisfy file-size budget, mypy ignore, SPEC update
CI feedback from PR #2563 surfaced three failures:

1. `file-size-budget`: parsers.py was 571 LOC, budget is 500. Split the
   per-language extractors into `_lang_python.py`, `_lang_js.py`,
   `_lang_rust.py`, `_lang_markdown.py` and a shared `_ts_common.py`
   helper module. `parsers.py` is now an 85-line dispatcher; the public
   API surface (`extract_python`, `extract_javascript`, ..., `dispatch`,
   `language_for`, `ParseResult`, `ParsedSymbol`) is preserved as
   re-exports so callers and tests don't change.

2. `quality-gate` mypy: the package uses untyped tree-sitter Node
   objects and optional-dep import guards (mcp, blake3, pathspec,
   watchdog) that mypy flags as unused. Add a `tool.mypy.overrides`
   entry for `codemap.*` matching the existing precedent for
   signal_toolkit widget mixins.

3. `Verify SPEC.md freshness`: bump to 1.1.131 with a Change Log entry
   describing the new package.

All 19 unit tests still pass after the split.
