+++ SPEC.md
@@ -27,8 +27,8 @@
 | **Primary Language(s)** | Python 3.11+, Rust, JavaScript, TypeScript |
 | **License**             | MIT                                        |
 | **Current Version**     | N/A                                        |
-| **Spec Version**        | 1.1.128                                    |
-| **Last Spec Update**    | 2026-05-07                                 |
+| **Spec Version**        | 1.1.131                                    |
+| **Last Spec Update**    | 2026-05-11                                 |
 
 ## 2. Purpose & Mission
 
@@ -506,6 +506,7 @@ Active development with stable core, continuous tool expansion, and web API in p
 
 | Date       | Version | Changes                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             |
 | ---------- | ------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| 2026-05-11 | 1.1.131 | Added shared `codemap` package (`src/shared/python/codemap/`) — tree-sitter symbol index over SQLite FTS5 with a 6-function pydantic query API (`search_code`, `get_symbol`, `who_calls`, `imports_of`, `neighbors`, `repo_summary`), CLI (`codemap rebuild/search/who-calls/export/info`), `watchdog` daemon (`codemap-watch`), and FastMCP server (`codemap-mcp`) so external coding agents inherit the same data the in-app chat consumes. `.codemap/` is gitignored; embedding layer deferred to a follow-up.                                                   |
 | 2026-05-11 | 1.1.130 | Hardened `signal_toolkit.calculus.Differentiator.differentiate` with an explicit `require(order >= 1, ...)` precondition so non-positive derivative orders raise `PreconditionError` instead of silently producing an empty derivative loop.                                                                                                                                                                                                                                                                                                                        |
 | 2026-05-11 | 1.1.129 | Added dynamic focus shifting to inline form validation within the Calculator app. This prevents keyboard focus traps by focusing the first invalid input (`.focus()`) and marking it with `aria-invalid="true"`.                                                                                                                                                                                                                                                                                                                                                    |
 | 2026-05-07 | 1.1.128 | Pre-compiled ODE Solver derivative expressions outside the RK4 loop while preserving the existing non-finite fallback behavior, so singular or overflowing user formulas still collapse to `0` instead of poisoning the integration state with `NaN` or `Infinity`.                                                                                                                                                                                                                                                                                                 |
