+++ src/shared/python/codemap/_lang_rust.py
@@ -1,0 +1,97 @@
+"""Rust tree-sitter extractor."""
+
+from __future__ import annotations
+
+from ._ts_common import (
+    ParsedSymbol,
+    ParseResult,
+    first_child,
+    get_parser,
+    line_range,
+    text_of,
+    to_bytes,
+)
+
+
+def _walk(node, source: bytes, prefix: str, out: list[ParsedSymbol]) -> None:
+    for child in node.children:
+        t = child.type
+        if t == "function_item":
+            name_node = first_child(child, "identifier")
+            if name_node is None:
+                continue
+            name = text_of(name_node, source)
+            qualified = f"{prefix}::{name}" if prefix else name
+            start, end = line_range(child)
+            out.append(
+                ParsedSymbol(
+                    kind="function",
+                    name=name,
+                    qualified=qualified,
+                    sig=text_of(child, source).splitlines()[0].rstrip(" {"),
+                    docstring="",
+                    start_line=start,
+                    end_line=end,
+                )
+            )
+        elif t == "struct_item":
+            name_node = first_child(child, "type_identifier")
+            if name_node is None:
+                continue
+            name = text_of(name_node, source)
+            qualified = f"{prefix}::{name}" if prefix else name
+            start, end = line_range(child)
+            out.append(
+                ParsedSymbol(
+                    kind="struct",
+                    name=name,
+                    qualified=qualified,
+                    sig=f"struct {name}",
+                    docstring="",
+                    start_line=start,
+                    end_line=end,
+                )
+            )
+        elif t == "impl_item":
+            type_node = first_child(child, "type_identifier")
+            sub_prefix = prefix
+            if type_node is not None:
+                sub_prefix = (
+                    f"{prefix}::{text_of(type_node, source)}"
+                    if prefix
+                    else text_of(type_node, source)
+                )
+            body = first_child(child, "declaration_list")
+            if body is not None:
+                _walk(body, source, sub_prefix, out)
+        elif t == "mod_item":
+            name_node = first_child(child, "identifier")
+            if name_node is None:
+                continue
+            name = text_of(name_node, source)
+            sub_prefix = f"{prefix}::{name}" if prefix else name
+            body = first_child(child, "declaration_list")
+            if body is not None:
+                _walk(body, source, sub_prefix, out)
+
+
+def _imports(root, source: bytes) -> list[str]:
+    out: list[str] = []
+    for c in root.children:
+        if c.type == "use_declaration":
+            out.append(text_of(c, source).strip().rstrip(";"))
+    return out
+
+
+def extract(path: str, source: str | bytes) -> ParseResult:
+    parser = get_parser("rust")
+    if parser is None:
+        return ParseResult("rust", [], [])
+    src = to_bytes(source)
+    tree = parser.parse(src)
+    symbols: list[ParsedSymbol] = []
+    _walk(tree.root_node, src, "", symbols)
+    return ParseResult("rust", _imports(tree.root_node, src), symbols)
+
+
+__all__ = ["extract"]
